requirements.txt
Flask
google-cloud-storage
google-generativeai
Pillow
Werkzeug
gunicorn